requirements.txt
fastapi==0.115.11
groq==0.19.0
groundx==2.2.3
pydantic==2.10.6
pyngrok==7.2.3
Requests==2.32.3
uvicorn==0.34.0
python-dotenv==1.0.1
